+++ source/rafcon/gui/start.py
@@ -132,22 +132,25 @@ def setup_argument_parser():
 
     :return: The parser object
     """
-    home_path = filesystem.get_home_path()
+    default_config_path = filesystem.get_default_config_path()
+    filesystem.create_path(default_config_path)
 
     parser = core_singletons.argument_parser
     parser.add_argument('-n', '--new', action='store_true', help=_("whether to create a new state-machine"))
     parser.add_argument('-o', '--open', action='store', nargs='*', type=parse_state_machine_path,
-                        dest='state_machine_paths', metavar='path', help=_(
-            "specify directories of state-machines that shall be opened. Paths must contain a statemachine.yaml file"))
+                        dest='state_machine_paths', metavar='path',
+                        help=_("specify directories of state-machines that shall be opened. "
+                               "Paths must contain a statemachine.yaml file"))
     parser.add_argument('-c', '--config', action='store', type=config_path, metavar='path', dest='config_path',
-                        default=home_path, nargs='?', const=home_path,
-                        help=_(
-                            "path to the configuration file config.yaml. Use 'None' to prevent the generation of a config file and use the default configuration. Default: {0}").format(
-                            home_path))
+                        default=default_config_path, nargs='?', const=default_config_path,
+                        help=_("path to the configuration file config.yaml. Use 'None' to prevent the generation of a "
+                               "config file and use the default configuration. Default: {0}"
+                               "").format(default_config_path))
     parser.add_argument('-g', '--gui_config', action='store', type=config_path, metavar='path', dest='gui_config_path',
-                        default=home_path, nargs='?', const=home_path, help=_(
-            "path to the configuration file gui_config.yaml. Use 'None' to prevent the generation of a config file and use the default configuration. Default: {0}").format(
-            home_path))
+                        default=default_config_path, nargs='?', const=default_config_path,
+                        help=_("path to the configuration file gui_config.yaml. Use 'None' to prevent the generation "
+                               "of a config file and use the default configuration. Default: {0}"
+                               "").format(default_config_path))
     parser.add_argument('-ss', '--start_state_machine', metavar='path', dest='start_state_machine_flag',
                         default=False, nargs='?',
                         help=_("a flag to specify if the state machine should be started after launching rafcon"))